palindromeString.c: simplify creation and output of the answer string
The `flag` variable has been removed. Instead, the `answer_str` variable added,
which points to the empty string or the "not " string, depending on whether the
input string is a palindrome or not.

The output string is generated during the `printf()` call using the
`answer_str` variable.

Also, the `if (flag)` operator and one of the `printf()` calls are unnecessary
now and has been removed.